+++ resources/assets/js/components/enso/vuedatatable/Pagination.vue
@@ -4,19 +4,21 @@
         role="navigation"
         aria-label="pagination">
         <a class="pagination-previous"
-            :disabled="page === 1"
+            :disabled="page === 1 || loading"
             @click="jumpTo(page - 1)">
             {{ i18n('Previous') }}
         </a>
         <a class="pagination-next"
-            :disabled="page === pages"
+            :disabled="page === pages || loading"
             @click="jumpTo(page + 1)">
             {{ i18n('Next') }}
         </a>
-        <ul class="pagination-list">
+        <ul class="pagination-list"
+            v-if="extended">
             <li>
                 <a class="pagination-link"
                     :class="{ 'is-current': page === 1 }"
+                    :disabled="loading"
                     @click="jumpTo(1)">
                     1
                 </a>
@@ -30,6 +32,7 @@
                 :key="i">
                 <a class="pagination-link"
                     :class="{ 'is-current': page === i }"
+                    :disabled="loading"
                     @click="jumpTo(i)">
                     {{ i }}
                 </a>
@@ -42,6 +45,7 @@
             <li v-if="pages > 1">
                 <a class="pagination-link"
                     :class="{ 'is-current': page === pages }"
+                    :disabled="loading"
                     @click="jumpTo(pages)">
                     {{ pages }}
                 </a>
@@ -57,6 +61,10 @@ export default {
     name: 'Pagination',
 
     props: {
+        loading: {
+            type: Boolean,
+            required: true,
+        },
         records: {
             type: Number,
             required: true,
@@ -67,6 +75,10 @@ export default {
         },
         length: {
             type: Number,
+            required: true,
+        },
+        extended: {
+            type: Boolean,
             required: true,
         },
         i18n: {
@@ -101,7 +113,11 @@ export default {
             }
 
             if (this.atEnd) {
-                pages.push(this.pages - 3, this.pages - 2, this.pages - 1);
+                if (pages > 4) {
+                    pages.push(this.pages - 3);
+                }
+
+                pages.push(this.pages - 2, this.pages - 1);
                 return pages;
             }